fix(storage): separate RFC-64 runtime CAS proof
